[v0.88][tools] Split pr_cmd_inline lifecycle tests by concern (Closes #1718)
