+++ adl/src/cli/tests/pr_cmd_inline/lifecycle/mod.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+mod closeout;
+mod diagnosis;
+mod start_ready;
